:truck: Refs. #614 -- move celery task decorator to general utils
